Renamed the singup interface to be the userRequestInterface close#7
