+++ server/utils/interfaces/index.ts
@@ -1,5 +1,5 @@
 import CustomRequest from './requestInterface';
-import UserInterface from './signupRequestInterface';
+import UserInterface from './userRequestInterface';
 import PayloadInterface from './payloadInterface';
 import UserValInterface from './userValInterface';
